requirements.txt
streamlit==1.32.0
langchain==0.1.4
faiss-cpu
python-dotenv
langchain-community==0.0.16
pypdf
numpy>=1.24.0
sentence-transformers
torch
transformers
huggingface_hub